+++ quantify_scheduler/enums.py
@@ -41,6 +41,7 @@ class BinMode(StrEnum):  # type: ignore
     APPEND = "append"
     AVERAGE = "average"
     FIRST = "first"
+    DISTRIBUTION = "distribution"
     # N.B. in principle it is possible to specify other behaviours for
     # BinMode such as `SUM` or `OVERWRITE` but these are not
     # currently supported by any backend.